test/EV-Service-Center-Full/services/finance-service/requirements.txt
# File: services/finance-service/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
psycopg2-binary==2.9.9
Flask-Migrate==4.0.7
python-dotenv==1.0.1
gunicorn==21.2.0
Flask-Cors==4.0.1
Werkzeug<3.0.0
requests==2.31.0
Flask-JWT-Extended==4.6.0